reverted codecept version to 3.7.4 to fix functional tests
This was required due to an error raised during functional tests when running under node 18 from the cucumber messages dependency. This can be fixed by either updating to a new version of node, or modyfying the dependency. As we seem to be using 18 in production still, I've gone with the dependency change, aligning it with other team projects.